% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{aDebug}
\alias{aDebug}
\title{aDebug}
\usage{
aDebug(...)
}
\arguments{
\item{...}{Una o varias cadena de texto.}
}
\value{
Nada, Imprime las cadenas que se el envían. Si DEBUG es TRUE
}
\description{
Envía varios textos de debug a la salida, concatenados en una sóla linea, cuando la variable DEBUG está establecida en TRUE.
}
\examples{
incremento <- 47
DEBUG <- TRUE
aDebug("Incremento: ", incremento, "Seg") #-> "Incremento: 47Seg"
}
\author{
Jose Alejandro Morán Pérez
}
