% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{aSalida}
\alias{aSalida}
\title{aSalida}
\usage{
aSalida(...)
}
\arguments{
\item{...}{Una o varias cadena de texto.}
}
\value{
Nada, Imprime las cadenas que se el envían
}
\description{
Envía varios textos a la salida, concatenados en una sóla linea.
}
\examples{
incremento <- 47
aSalida("Incremento: ", incremento, "Seg") #-> "Incremento: 47Seg"
}
\author{
Jose Alejandro Morán Pérez
}
