% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{asegurarExistencia}
\alias{asegurarExistencia}
\title{asegurarExistencia}
\usage{
asegurarExistencia(archivo, nombre = archivo)
}
\arguments{
\item{archivo}{La ruta completa del archivo}

\item{nombre}{El nombre con el que se quiere referir al archivo, si se omite se usará la ruta completa}
}
\value{
TRUE o FALSE según exista, o no, el archivo.
}
\description{
Comprueba si un archivo exixte en el sistema operativo o no.
}
\details{
Si no existe, informa de ello por consola. Si existe, informa por DEBUG.
}
\examples{
asegurarExistencia(nombre_archivo_csv, "ImportanTe")
#-> TRUE (Debug: "OK: se ha encontrado el archivo ImportanTe")
}
\author{
Jose Alejandro Morán Pérez
}
