% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAlpha.R
\name{bwt}
\alias{bwt}
\title{Realiza la transformación de Burrows-Wheeler}
\source{
https://gist.github.com/aaronwolen/cbdd74180a714267bf0c
}
\usage{
bwt(x, eof = "!")
}
\arguments{
\item{x}{El texto a transformar}

\item{eof}{Carácter de control para marcar el número de línea que puede reconstruir el texto transformado. No puede estar contenido en el texto. Por defecto, !}
}
\description{
Realiza la transformación de Burrows-Wheeler
}
\examples{
bwt('SIX.MIXED.PIXIES.SIFT') #-> "TXDSEXIIXSSMP..E.!FIII"
}
\seealso{
ibwt La operación contraria
}
