% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasDependientes.R
\name{color_transparente}
\alias{color_transparente}
\title{color_transparente}
\usage{
color_transparente(color, percent = 50, name = NULL)
}
\arguments{
\item{color}{El nombre del color}

\item{percent}{Porcentaje de transparencia}

\item{name}{Un nombre, opcional, para el color}
}
\description{
Crea un color transparente
}
\examples{
color_transparente("#11aaff", 25) #-> "#11AAFFBF"
}
\author{
Jose Alejandro Morán Pérez
}
