% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasDependientes.R
\name{crearTituloExcel}
\alias{crearTituloExcel}
\title{crearTituloExcel}
\usage{
crearTituloExcel(hoja, fila, columna, texto, estilo)
}
\arguments{
\item{hoja}{El objeto de hoja de excel en que poner el título}

\item{fila}{Coordenada de la fila de la celda en que situar el título}

\item{columna}{Coordenada de la columna de la celda en que situar el título}

\item{texto}{Texto del título. Si es nulo, no se establece el texto, sólo se cambia el formato.}

\item{estilo}{Estilo en que irá el título.}
}
\description{
Crea un título formateado en una celda de una hoja de excel
}
\examples{
crearTituloExcel(hoja = sheet, fila = 2, columna = 2, estilo = TITLE_STYLE)
#-> ### Crea un título formateado en la celda B2
}
\author{
Jose Alejandro Morán Pérez
}
