% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasDependientes.R
\name{cualEsMiIP}
\alias{cualEsMiIP}
\title{cualEsMiIP}
\usage{
cualEsMiIP()
}
\value{
Una cadena con la IP asignada al ordenador.
}
\description{
Devuelve la ip actual del ordenador.
}
\examples{
cualEsMiIP() #-> "193.146.96.2"
}
\seealso{
\url{https://www.ipify.org/}
}
\author{
Jose Alejandro Morán Pérez
}
