% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{cuidado}
\alias{cuidado}
\title{cuidado}
\usage{
cuidado(...)
}
\arguments{
\item{...}{Una o varias cadena de texto.}
}
\value{
La cadena de peligro.
}
\description{
Envía varios textos al registro de warnings, concatenados en una sóla linea.
}
\examples{
 \dontrun{
numero <- 47
cuidado("Peligro: ", numero, "Seg") #-> "Peligro: 47Seg"
# "Warning message:
# "In cuidado("Peligro: ", numero, "Seg") : Peligro: 47Seg
}
}
\author{
Jose Alejandro Morán Pérez
}
