% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasDependientes.R
\name{desconectarTodasLasBasesDeDatosMysql}
\alias{desconectarTodasLasBasesDeDatosMysql}
\title{desconectarTodasLasBasesDeDatosMysql}
\usage{
desconectarTodasLasBasesDeDatosMysql()
}
\description{
Cierra todas la conexiónes con servidores Mysql activas
}
\examples{
 \dontrun{
desconectarTodasLasBasesDeDatosMysql()
}
}
\author{
Jose Alejandro Morán Pérez
}
