% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{existe}
\alias{existe}
\title{existe}
\usage{
existe(objeto)
}
\arguments{
\item{objeto}{El nombre del objeto}
}
\value{
TRUE si existe, FALSE si no.
}
\description{
Comprueba si un objeto existe en memoria
}
\examples{
temporal3x3mpl0 <- "un objeto"
existe(temporal3x3mpl0) #-> TRUE
existe("temporal3x3mpl0") #-> TRUE
rm(list = "temporal3x3mpl0")
existe("temporal3x3mpl0") #-> FALSE
}
\author{
Jose Alejandro Morán Pérez
}
