% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{extraerFecha}
\alias{extraerFecha}
\title{extraerFecha}
\usage{
extraerFecha(cadena)
}
\arguments{
\item{cadena}{La cadena de fecha y hora.}
}
\value{
La subcadena de la fecha.Si el formato es incorrecto devuelve una cadena vacía y lanza un aviso.
}
\description{
Extrae la fecha de una cadena que tiene fecha y hora
}
\examples{
extraerFecha("2019-10-02 23:33:14") #-> "2019-10-02"
}
\author{
Jose Alejandro Morán Pérez
}
