% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{extraerHora}
\alias{extraerHora}
\title{extraerHora}
\usage{
extraerHora(cadena)
}
\arguments{
\item{cadena}{La cadena de fecha y hora.}
}
\value{
La subcadena de la hora.Si el formato es incorrecto devuelve una cadena vacía y lanza un aviso.
}
\description{
Extrae la hora de una cadena que tiene fecha y hora
}
\examples{
extraerHora("2019-10-02 23:33:14") #-> "23:33:14"
}
\author{
Jose Alejandro Morán Pérez
}
