% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{format.segundos}
\alias{format.segundos}
\title{format.segundos}
\usage{
\method{format}{segundos}(start_time)
}
\arguments{
\item{start_time}{Un número de segundos.}
}
\value{
Una cadena de texto en formato \code{h:mm:ss}.
}
\description{
Formatea un dato de segundos para que mostrar las horas y minutos que representan.
}
\examples{
format.segundos(605) #| Produce:0:10:05
}
\author{
Jose Alejandro Morán Pérez
}
