% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{getTipo}
\alias{getTipo}
\title{getTipo}
\usage{
getTipo(nombreDeFichero)
}
\arguments{
\item{nombreDeFichero}{La cadena del nombre del fichero.}
}
\value{
La subcadena de la extensión, o "" si no encuentra el punto..
}
\description{
Devuelve el tipo de archivo (extensión) a partir de su nombre
}
\examples{
getTipo("00176a223d658759746323cc1281e93d.jpg") #-> "jpg"
}
\author{
Jose Alejandro Morán Pérez
}
