% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasDependientes.R
\name{guardar_tsv_utf8_con_bom}
\alias{guardar_tsv_utf8_con_bom}
\title{guardar_tsv_utf8_con_bom}
\usage{
guardar_tsv_utf8_con_bom(datos, nombre_fichero)
}
\arguments{
\item{datos}{El conjunto de datos a guardar.}

\item{nombre_fichero}{El nombre del fichero con que se guardarán los datos.}
}
\value{
nada
}
\description{
Guarda un dataframe con fommato de archivo .tsv codificado en UTF-8 con BOM
}
\examples{
 \dontrun{
guardar_tsv_utf8_con_bom(datos, "datos.tsv")
}
}
\author{
Jose Alejandro Morán Pérez
}
