% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{haSalida}
\alias{haSalida}
\title{haSalida}
\usage{
haSalida(...)
}
\arguments{
\item{...}{Una o varias cadena de texto.}
}
\value{
Nada, Imprime las cadenas que se el envían
}
\description{
Envía varios textos a la salida, concatenados en una sóla linea, embebidos entre tag's P de html. Es adecuada para mostrar resultados 'asis' en documentos rmarkdown.
}
\examples{
incremento <- 47
haSalida("Incremento: ", incremento, "Seg") #-> "<p>Incremento: 47Seg<p>"
}
\author{
Jose Alejandro Morán Pérez
}
