% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{hay}
\alias{hay}
\title{hay}
\usage{
hay(tabla)
}
\arguments{
\item{tabla}{Una tabla de datos medible.}
}
\value{
TRUE o FALSE según haya, o no, registros en la tabla.
}
\description{
Indica si una tabla contiene datos
}
\examples{
hay(c(3, 3)) #-> TRUE
}
\author{
Jose Alejandro Morán Pérez
}
