% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAle.R
\docType{package}
\name{HerramientasAle}
\alias{HerramientasAle}
\title{HerramientasAle}
\description{
Un paquete de herramientas y utilidades para facilitar la creación de scripts en R
}
\details{
El paquete proporciona utlidades como
aSalida, tamagno, ...
}
