% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{imprimir}
\alias{imprimir}
\title{imprimir (Deprecado)}
\usage{
imprimir(datos, sustituto = "")
}
\arguments{
\item{datos}{Un conjunto de datos que deben imprimir. Se esperan dataframes}

\item{sustituto}{El valor por el que se reemplazan los NA. Por defecto: vacío}
}
\description{
Imprime un conjunto de datos sustituyendo los NA por otro valor
}
\details{
Deprecado: Esta función va a ser eliminada en próximas versiones. Su nombre es demasiado genérico. Mejor usar imprimirNAs
}
\author{
Jose Alejandro Morán Pérez
}
