% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{imprimirNAs}
\alias{imprimirNAs}
\title{imprimirNAs}
\usage{
imprimirNAs(datos, sustituto = "")
}
\arguments{
\item{datos}{Un conjunto de datos que deben imprimir. Se esperan dataframes}

\item{sustituto}{El valor por el que se reemplazan los NA. Por defecto: vacío}
}
\description{
Imprime un conjunto de datos sustituyendo los NA por otro valor
}
\author{
Jose Alejandro Morán Pérez
}
