% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{insertarTitulo}
\alias{insertarTitulo}
\title{insertarTitulo}
\usage{
insertarTitulo(titulo, nivel)
}
\arguments{
\item{titulo}{El título a inserta.}

\item{nivel}{El nivel que debe tener dicho título.}
}
\description{
Inserta un título para R Markdown del nivel que se necesite
}
\examples{
insertarTitulo("Título de nivel 3", 3) #-> ### Título de nivel 3
}
\author{
Jose Alejandro Morán Pérez
}
