% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{left}
\alias{left}
\title{left}
\source{
https://www.rforexcelusers.com/how-to-mid-right-left-r/
}
\usage{
left(text, num_char)
}
\arguments{
\item{text}{La cadena de caracteres.}

\item{num_char}{Número de caracteres.}
}
\description{
Extrae la parte izquierda de una cadena de caracteres.
}
\author{
Jose Alejandro Morán Pérez
}
