% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAlpha.R
\name{limpiarDatosDeHoja}
\alias{limpiarDatosDeHoja}
\title{limpiarDatosDeHoja}
\usage{
limpiarDatosDeHoja(datos, titulos = NULL)
}
\arguments{
\item{datos}{Los datos a limpiar.}

\item{titulos}{Los títulos a añadir al resultado. Si no se incluyen, se extraen de los datos}
}
\value{
un dataframe con los datos límpios, el título sin puntos y con, al menos, una fila y 2 columnas.
}
\description{
Elimina las filas vacías de los datos extraídos de una hoja de excel.
}
\details{
Sustituye los puntos por espacios en el título.
Si los datos no los tienen, añade, al menos una fila y 2 columnas, para que se puedan dibujar como tablas
Está pensado para eliminar filas inútiles de las hojas de datos complejas.
}
\author{
Jose Alejandro Morán Pérez <jaleomp@gmail.com>
}
