% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{mid}
\alias{mid}
\title{mid}
\source{
https://www.rforexcelusers.com/how-to-mid-right-left-r/
}
\usage{
mid(text, start_num, num_char)
}
\arguments{
\item{text}{La cadena de caracteres}

\item{start_num}{La posición inicial de la extracción.}

\item{num_char}{Número de caracteres}
}
\description{
Extrae la parte central de una cadena de caracteres
}
\author{
Jose Alejandro Morán Pérez
}
