% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{recuento}
\alias{recuento}
\title{recuento}
\usage{
recuento(tabla)
}
\arguments{
\item{tabla}{Una tabla de datos medible.}
}
\value{
El número de registros en la tabla.
}
\description{
Cuenta el número de filas que tiene una tabla
}
\examples{
recuento(c(3, 3)) #-> 2
}
\author{
Jose Alejandro Morán Pérez
}
