% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{reemplazar.nas}
\alias{reemplazar.nas}
\title{reemplazar.nas}
\usage{
reemplazar.nas(data)
}
\arguments{
\item{data}{El dataframe}
}
\value{
El dataframe sin NAs
}
\description{
Cambia los NAs de un dataframe por espacios vacíos
}
\examples{
ejemplo <- iris[1:2, 1:2]
#     Sepal.Length Sepal.Width
#              5.1         3.5
#              4.9         3.0
ejemplo[2, 2] <- NA
#     Sepal.Length Sepal.Width
#              5.1         3.5
#              4.9         NA
reemplazar.nas(ejemplo) #->
#     Sepal.Length Sepal.Width
#              5.1         3.5
#              4.9
}
\author{
Jose Alejandro Morán Pérez
}
