% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{right}
\alias{right}
\title{right}
\source{
https://www.rforexcelusers.com/how-to-mid-right-left-r/
}
\usage{
right(text, num_char)
}
\arguments{
\item{text}{La cadena de caracteres.}

\item{num_char}{Número de caracteres.}
}
\description{
Extrae la parte derecha de una cadena de caracteres
}
\author{
Jose Alejandro Morán Pérez
}
