% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{sumar_fechas}
\alias{sumar_fechas}
\title{sumar_fechas}
\usage{
sumar_fechas(fechas, dias = 0, horas = 0, minutos = 0, segundos = 0)
}
\arguments{
\item{fechas}{la lista de fechas a modificar.}

\item{dias}{el número de días a añadir (opcional).}

\item{horas}{el número de horas a añadir (opcional).}

\item{minutos}{el número de minutos a añadir (opcional).}

\item{segundos}{el número de segundos a añadir (opcional).}
}
\value{
las fechas modificadas. Si no se puede convertir a fecha, lo mismo que se envíe.
}
\description{
Suma una cantidad de tiempo a una, o varias fechas. Adminte números negativos, para restar.
}
\examples{
sumar_fechas("2020-12-04 14:46:47", -4, 4) #-> "2020-11-30 18:46:47"
}
\author{
Jose Alejandro Morán Pérez
}
