% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{sustituir}
\alias{sustituir}
\title{sustituir (Deprecado)}
\usage{
sustituir(datos, sustituto = "·")
}
\arguments{
\item{datos}{Un conjunto de datos que deben imprimir. Se esperan dataframes.}

\item{sustituto}{El valor por el que se reemplazan los NA. Por defecto: ·.}
}
\value{
El conjunto de datos con los NAs sustituidos.
}
\description{
Sustituye los NA por otro valor en un conjunto de datos
}
\details{
Deprecado: Esta función va a ser eliminada en próximas versiones. Su nombre es demasiado genérico. Mejor usar sustituirNAs
}
\examples{
ejemplo <- iris[1:2, 1:2]
ejemplo
# Sepal.Length Sepal.Width
#          5.1         3.5
#          4.9         3.0
ejemplo[2, 2] <- NA
ejemplo
# Sepal.Length Sepal.Width
#          5.1         3.5
#          4.9          NA
sustituir(ejemplo, "0")
# Sepal.Length Sepal.Width
#          5.1         3.5
#          4.9           0
}
\references{
https://stackoverflow.com/questions/19516302/do-not-print-na-when-printing-data-frame
}
\seealso{
sustituirNAs
}
\author{
Jose Alejandro Morán Pérez
}
