% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{tamagno}
\alias{tamagno}
\title{tamagno}
\usage{
tamagno(objeto)
}
\arguments{
\item{objeto}{El objeto a evaluar.}
}
\value{
El número de filas y columnas que tiene. O el número de elementos, si es una lista.
}
\description{
Calcula el tamaño de un objeto.
}
\examples{
listado <- rep("a", 256)
tamagno(listado) #-> 256
}
\author{
Jose Alejandro Morán Pérez
}
