% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAlpha.R
\name{trataDeComprimir}
\alias{trataDeComprimir}
\title{Trata de comprimir un texto en una cadena lineal que mantiene toda la información y puede ser descomprimido.}
\usage{
trataDeComprimir(
  texto_original,
  aglutinador = "·",
  separador = "ç",
  eof = "!",
  forzar = FALSE
)
}
\arguments{
\item{texto_original}{El texto que se quiere comprimir.}

\item{aglutinador}{Caracter de relleno que separa los números. No debería ser un número ni estar contenido en el texto. Por defecto ·.}

\item{separador}{Caracter que separa las 2 partes de la cadena se salida. No debería estar contenido en el texto. Por defecto, ç.}

\item{eof}{Marca de posición de fila correcta para el algoritmo de transformación de Burrows-Wheeler. No debería estar contenido en el texto. Por defecto, !.}

\item{forzar}{Indica si se debe forzar la conversión del texto. Si es TRUE, se devolverá el resultado del algoritmo incluso aunque el texto resultante sea más largo que el original. Puede servir para encriptarlo a humanos. Por defecto, FALSE.}
}
\description{
Es útil para textos repetitivos. No debería usarse con cadenas excesivamente grandes
El texto original no puede contener ni el carácter aglutinador, ni el separador ni el eof.
No se garantiza que se haga la transformación salvo que forzar sea = TRUE.
En caso contrario, sólo se devuelve el texto transformado si es menor que el origina.
}
\examples{
# example code
trataDeComprimir("En un lugar de la mancha", forzar = TRUE)
#-> "1·1·1·1·1·1·1·1·1·1·1·1·1·1·1·1·3·1·1·2·1·1çrenanalhmgn d!uc uEal "
}
\seealso{
bwt

trataDeDescomprimir
}
\author{
Jose Alejandro Morán Pérez <jaleomp@gmail.com>
}
