% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAlpha.R
\name{trataDeDescomprimir}
\alias{trataDeDescomprimir}
\title{Trata de descomprimir la cadena producida por trataDeComprimir.}
\usage{
trataDeDescomprimir(
  texto_comprimido,
  aglutinador = "·",
  separador = "ç",
  eof = "!"
)
}
\arguments{
\item{texto_comprimido}{El texto que se quiere descomprimir.}

\item{aglutinador}{Caracter de relleno que separa los números. Por defecto ·.}

\item{separador}{Caracter que separa las 2 partes de la cadena se salida.  Por defecto, ç.}

\item{eof}{Marca de posición de fila correcta para el algoritmo de transformación de Burrows-Wheeler. Por defecto, !.}
}
\description{
Es útil para textos repetitivos. No debería usarse con cadenas excesivamente grandes. Es muy lento.
Los parámetros de descompresión deben ser exactamente iguales a los parámetros con que fue producida la cadena. Especialmente el carácter aglutinador, el separador y el eof.
Si no ha sido comprimido con trataDeComprimir, se devuelve el mismo texto.
}
\examples{
# example code
 trataDeDescomprimir("1·1·1·1·1·1·1·1·1·1·1·1·1·1·1·1·3·1·1·2·1·1çrenanalhmgn d!uc uEal ")
 # -> "En un lugar de la mancha"
}
\seealso{
ibwt

trataDeComprimir
}
\author{
Jose Alejandro Morán Pérez <jaleomp@gmail.com>
}
