% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasDependientes.R
\name{extraerRutaDeWindows}
\alias{extraerRutaDeWindows}
\title{extraerRutaDeWindows}
\usage{
extraerRutaDeWindows(path = "clipboard", copiarAPortaPapeles = TRUE)
}
\arguments{
\item{path}{El texto de la variable que se quiere convertir. Por defecto, la cadena almacenada en el portapapeles.}

\item{copiarAPortaPapeles}{Indica si copiar, o no, el resultado al portapapeles. Por defecto, si.}
}
\value{
Una ruta compatible con R.
}
\description{
Formatéa una ruta del sistema operativo Windows para que sea posible introducirla en un script de R sin necesidad e hacer cambios adicionales.
}
\details{
Esta utilidad tiene un comportamiento especial: utiliza el contendio del portapapeles, por tanto, la ruta del dirctorio debe estar en el portapapeles.
Al ejecutrar esta función, se imprime la ruta en el formato que se necesita
}
\examples{
extraerRutaDeWindows() #-> "G:/Unidades compartidas/EMPRESA/BBDD/JSON"
}
\author{
Jose Alejandro Morán Pérez
}
