% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerramientasAutonomas.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Not in}
\source{
\url{http://stackoverflow.com/questions/5831794/opposite-of-in}
}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{uno de los operandos.}

\item{y}{otro operando. Con quien se compara el primero.}
}
\description{
Operador de contrario a \%in\%
}
\examples{
a <- c(1, 2, 3, 4, 5)
b <- c(1, 2, 3, 5, 7)
a \%!in\% b
}
